Moved construct_circ_avg_image to core.image
